% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce10_ARCC}
\alias{ce10_ARCC}
\title{Interactions identified in L3 C. elegans by ARC-C}
\format{
\subsection{\code{ce10_ARCC}}{
}
}
\source{
\url{https://genome.cshlp.org/content/early/2021/12/21/gr.275669.121}
}
\usage{
ce10_ARCC
}
\value{
A \code{GInteractions} object
}
\description{
Supplemental Table 2 obtained from Genome Biology online publication.

Huang N, Seow WQ, Appert A, Dong Y, Stempor P and Ahringer J
Accessible Region Conformation Capture (ARC-C) gives high-resolution
insights into genome architecture and regulation.
Genome Res 2022 Feb;32(2):357-366. PMID: 34933938
}
\keyword{internal}
