% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce10_REs}
\alias{ce10_REs}
\title{Annotated regulatory elements in C. elegans}
\format{
\subsection{\code{ce10_REs}}{
}
}
\source{
\url{https://genome.cshlp.org/content/early/2021/12/21/gr.275669.121}
}
\usage{
ce10_REs
}
\value{
A \code{GRanges} object
}
\description{
Figure 2 - Source data 1 obtained from eLife online publication.

Jänes J, Dong Y, Schoof M, Serizay J, Appert A, Cerrato C, Woodbury C,
Chen R, Gemma C, Huang N, Kissiov D, Stempor P, Steward A, Zeiser E,
Sauer S and Ahringer J
Chromatin accessibility dynamics across C. elegansdevelopment and ageing.
Elife 2018 Oct 26;7. PMID: 30362940
}
\keyword{internal}
