% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-mutate.R
\name{mutate.GInteractions}
\alias{mutate.GInteractions}
\title{Mutate columns from a GInteractions object}
\usage{
\method{mutate}{GInteractions}(.data, ...)
}
\arguments{
\item{.data}{a GInteractions object}

\item{...}{Optional named arguments specifying which the columns in .data
to create/modify.}
}
\value{
a GInteractions object.
}
\description{
Mutate columns from a GInteractions object
}
\examples{
gi <- read.table(text = "
chr1 10 20 chr1 50 51
chr1 10 50 chr1 30 40",
col.names = c("chr1", "start1", "end1", "chr2", "start2", "end2")) |> 
  as_ginteractions(seqnames1 = chr1, seqnames2 = chr2)
  
gi |> 
  mutate(type = 'cis', score = runif(2)) |> 
  mutate(type2 = type)
}
