% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginteractions-setters.R
\name{ginteractions-setters}
\alias{ginteractions-setters}
\alias{first<-,GInteractions-method}
\alias{second<-,GInteractions-method}
\alias{set_seqnames1}
\alias{set_seqnames2}
\alias{set_start1}
\alias{set_start2}
\alias{set_end1}
\alias{set_end2}
\alias{set_width1}
\alias{set_width2}
\alias{set_strand1}
\alias{set_strand2}
\alias{set_seqnames1,GInteractions,factor-method}
\alias{set_seqnames2,GInteractions,factor-method}
\alias{set_start1,GInteractions,numeric-method}
\alias{set_start2,GInteractions,numeric-method}
\alias{set_end1,GInteractions,numeric-method}
\alias{set_end2,GInteractions,numeric-method}
\alias{set_width1,GInteractions,numeric-method}
\alias{set_width2,GInteractions,numeric-method}
\alias{set_width1,AnchoredPinnedGInteractions,numeric-method}
\alias{set_width2,AnchoredPinnedGInteractions,numeric-method}
\alias{set_strand1,GInteractions,character-method}
\alias{set_strand2,GInteractions,character-method}
\title{Internal GInteractions setters}
\usage{
\S4method{first}{GInteractions}(x) <- value

\S4method{second}{GInteractions}(x) <- value

set_seqnames1(x, value)

set_seqnames2(x, value)

set_start1(x, value)

set_start2(x, value)

set_end1(x, value)

set_end2(x, value)

set_width1(x, value)

set_width2(x, value)

set_strand1(x, value)

set_strand2(x, value)

\S4method{set_seqnames1}{GInteractions,factor}(x, value)

\S4method{set_seqnames2}{GInteractions,factor}(x, value)

\S4method{set_start1}{GInteractions,numeric}(x, value)

\S4method{set_start2}{GInteractions,numeric}(x, value)

\S4method{set_end1}{GInteractions,numeric}(x, value)

\S4method{set_end2}{GInteractions,numeric}(x, value)

\S4method{set_width1}{GInteractions,numeric}(x, value)

\S4method{set_width2}{GInteractions,numeric}(x, value)

\S4method{set_width1}{AnchoredPinnedGInteractions,numeric}(x, value)

\S4method{set_width2}{AnchoredPinnedGInteractions,numeric}(x, value)

\S4method{set_strand1}{GInteractions,character}(x, value)

\S4method{set_strand2}{GInteractions,character}(x, value)
}
\arguments{
\item{x}{a GInteractions object}

\item{value}{a value passed to the corresponding field}
}
\value{
A modified GInteractions
}
\description{
Internal GInteractions setters
}
\keyword{internal}
