% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GM12878_HiCCUPS}
\alias{GM12878_HiCCUPS}
\title{Loops identified in GM12878 with HiCCUPS}
\format{
An object of class \code{GInteractions} of length 9448.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
GM12878_HiCCUPS
}
\value{
A \code{GInteractions} object
}
\description{
File obtained from GEO entry \code{GSE63525}
(GSE63525_GM12878_primary+replicate_HiCCUPS_looplist.txt.gz).

Rao SS, Huntley MH, Durand NC, Stamenova EK et al.
A 3D map of the human genome at kilobase resolution
reveals principles of chromatin looping.
Cell 2014 Dec 18;159(7):1665-80. PMID: 25497547
}
\keyword{internal}
