% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plyinteractions-data}
\alias{plyinteractions-data}
\alias{GM12878_HiCCUPS}
\alias{ce10_ARCC}
\alias{ce10_REs}
\title{Data files provided in the plyinteractions package}
\format{
A \code{GInteractions} object

A \code{GInteractions} object

A \code{GInteractions} object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}

\url{https://genome.cshlp.org/content/early/2021/12/21/gr.275669.121}

\url{https://genome.cshlp.org/content/early/2021/12/21/gr.275669.121}
}
\usage{
data(GM12878_HiCCUPS)

data(GM12878_HiCCUPS)

data(GM12878_HiCCUPS)
}
\description{
\itemize{
\item Loops identified in GM12878 with HiCCUPS:
File obtained from GEO entry \code{GSE63525}
(GSE63525_GM12878_primary+replicate_HiCCUPS_looplist.txt.gz).
Rao SS, Huntley MH, Durand NC, Stamenova EK et al.
A 3D map of the human genome at kilobase resolution
reveals principles of chromatin looping.
Cell 2014 Dec 18;159(7):1665-80. PMID: 25497547
\item Interactions identified in L3 C. elegans by ARC-C:
Supplemental Table 2 obtained from Genome Biology online publication.
Huang N, Seow WQ, Appert A, Dong Y, Stempor P and Ahringer J
Accessible Region Conformation Capture (ARC-C) gives high-resolution
insights into genome architecture and regulation.
Genome Res 2022 Feb;32(2):357-366. PMID: 34933938
\item Annotated regulatory elements in C. elegans:
Figure 2 - Source data 1 obtained from eLife online publication.
Jänes J, Dong Y, Schoof M, Serizay J, Appert A, Cerrato C, Woodbury C,
Chen R, Gemma C, Huang N, Kissiov D, Stempor P, Steward A, Zeiser E,
Sauer S and Ahringer J
Chromatin accessibility dynamics across C. elegansdevelopment and ageing.
Elife 2018 Oct 26;7. PMID: 30362940
}
}
\keyword{datasets}
