% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-dplyr.R, R/reexports-plyranges.R,
%   R/reexports.R
\name{mutate}
\alias{mutate}
\alias{slice}
\alias{select}
\alias{filter}
\alias{arrange}
\alias{rename}
\alias{group_by}
\alias{group_data}
\alias{group_keys}
\alias{group_indices}
\alias{group_vars}
\alias{groups}
\alias{group_size}
\alias{n_groups}
\alias{group_rows}
\alias{ungroup}
\alias{summarize}
\alias{summarise}
\alias{tally}
\alias{count}
\alias{anchor}
\alias{unanchor}
\alias{anchor_start}
\alias{anchor_end}
\alias{anchor_center}
\alias{anchor_3p}
\alias{anchor_5p}
\alias{stretch}
\alias{find_overlaps}
\alias{find_overlaps_directed}
\alias{count_overlaps}
\alias{count_overlaps_directed}
\alias{filter_by_overlaps}
\alias{filter_by_non_overlaps}
\alias{join_overlap_left}
\alias{join_overlap_left_directed}
\alias{reexports}
\title{reexports}
\usage{
mutate(.data, ...)

slice(.data, ..., .by = NULL, .preserve = FALSE)

select(.data, ...)

filter(.data, ..., .by = NULL, .preserve = FALSE)

arrange(.data, ..., .by_group = FALSE)

rename(.data, ...)

group_by(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))

group_data(.data)

group_keys(.tbl, ...)

group_indices(.data, ...)

group_vars(x)

groups(x)

group_size(x)

n_groups(x)

group_rows(.data)

ungroup(x, ...)

summarize(.data, ..., .by = NULL, .groups = NULL)

summarise(.data, ..., .by = NULL, .groups = NULL)

tally(x, wt = NULL, sort = FALSE, name = NULL)

count(x, ..., wt = NULL, sort = FALSE, name = NULL)

anchor(x)

unanchor(x)

anchor_start(x)

anchor_end(x)

anchor_center(x)

anchor_3p(x)

anchor_5p(x)

stretch(x, extend)

find_overlaps(x, y, maxgap, minoverlap, suffix = c(".x", ".y"))

find_overlaps_directed(x, y, maxgap, minoverlap, suffix = c(".x", ".y"))

count_overlaps(x, y, maxgap, minoverlap)

count_overlaps_directed(x, y, maxgap, minoverlap)

filter_by_overlaps(x, y, maxgap = -1L, minoverlap = 0L)

filter_by_non_overlaps(x, y, maxgap, minoverlap)

join_overlap_left(x, y, maxgap, minoverlap, suffix = c(".x", ".y"))

join_overlap_left_directed(x, y, maxgap, minoverlap, suffix = c(".x", ".y"))
}
\value{
Depending on the re-exported function
}
\description{
reexports
}
\examples{
1 + 1
}
\keyword{internal}
