% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show-GInteractions}
\alias{show-GInteractions}
\alias{show,GInteractions-method}
\alias{show,AnchoredPinnedGInteractions-method}
\alias{show,GroupedGInteractions-method}
\alias{show,PinnedGInteractions-method}
\title{show method for \code{GInteractions} objects}
\arguments{
\item{object}{a \verb{(Anchored/Pinned/Grouped)GInteractions} object}
}
\value{
\verb{Prints a message to the console describing the contents of a }GInteractions` object.
}
\description{
show method for \code{GInteractions} objects
}
\examples{
pairsf <- system.file('extdata', 'pairs.gz', package = 'plyinteractions')
pairs <- read.table(pairsf, comment.char = '#', header = FALSE)
pairs |> 
  as_ginteractions(
    seqnames1 = V2, start1 = V3, width1 = 1, strand1 = V6, 
    seqnames2 = V4, start2 = V5, width2 = 1, strand2 = V7,
    starts.in.df.are.0based = TRUE
  )
}
