% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-DelegatingGInteractions.R
\name{delegating-ginteractions-methods}
\alias{delegating-ginteractions-methods}
\alias{anchors1,DelegatingGInteractions-method}
\alias{ranges1,DelegatingGInteractions-method}
\alias{seqnames1,DelegatingGInteractions-method}
\alias{start1,DelegatingGInteractions-method}
\alias{end1,DelegatingGInteractions-method}
\alias{width1,DelegatingGInteractions-method}
\alias{strand1,DelegatingGInteractions-method}
\alias{anchors2,DelegatingGInteractions-method}
\alias{ranges2,DelegatingGInteractions-method}
\alias{seqnames2,DelegatingGInteractions-method}
\alias{start2,DelegatingGInteractions-method}
\alias{end2,DelegatingGInteractions-method}
\alias{width2,DelegatingGInteractions-method}
\alias{strand2,DelegatingGInteractions-method}
\alias{anchors,DelegatingGInteractions-method}
\alias{regions,DelegatingGInteractions-method}
\alias{seqinfo,DelegatingGInteractions-method}
\alias{mcols,DelegatingGInteractions-method}
\alias{show,DelegatingGInteractions-method}
\title{Methods for DelegatingGInteractions objects}
\usage{
\S4method{anchors1}{DelegatingGInteractions}(x)

\S4method{ranges1}{DelegatingGInteractions}(x)

\S4method{seqnames1}{DelegatingGInteractions}(x)

\S4method{start1}{DelegatingGInteractions}(x)

\S4method{end1}{DelegatingGInteractions}(x)

\S4method{width1}{DelegatingGInteractions}(x)

\S4method{strand1}{DelegatingGInteractions}(x)

\S4method{anchors2}{DelegatingGInteractions}(x)

\S4method{ranges2}{DelegatingGInteractions}(x)

\S4method{seqnames2}{DelegatingGInteractions}(x)

\S4method{start2}{DelegatingGInteractions}(x)

\S4method{end2}{DelegatingGInteractions}(x)

\S4method{width2}{DelegatingGInteractions}(x)

\S4method{strand2}{DelegatingGInteractions}(x)

\S4method{anchors}{DelegatingGInteractions}(x)

\S4method{regions}{DelegatingGInteractions}(x)

\S4method{seqinfo}{DelegatingGInteractions}(x)

\S4method{mcols}{DelegatingGInteractions}(x)

\S4method{show}{DelegatingGInteractions}(object)
}
\description{
Methods for DelegatingGInteractions objects
}
\keyword{internal}
