% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{ginteractions-annotate}
\alias{ginteractions-annotate}
\alias{annotate,GInteractions,GRanges,character-method}
\title{Annotate both anchors of a GInteractions}
\usage{
annotate(x, y, label, ...)

\S4method{annotate}{GInteractions,GRanges,character}(x, y, label)
}
\arguments{
\item{x}{a GInteractions object}

\item{y}{a GRanges object to extract annotations from}

\item{label}{Column name from \code{y} to use to extract annotations}
}
\value{
a GInteractions object with two extra metadata columns named
\code{annotation_first} and \code{annotation_second}.
}
\description{
Annotate both anchors of a GInteractions
}
\examples{
data(GM12878_HiCCUPS, package = 'plyinteractions')
annots <- AnnotationHub::AnnotationHub()[['AH75190']]
annots <- lapply(annots, function(x) {seqlevelsStyle(x) <- "NCBI" ; x})

annotate(GM12878_HiCCUPS, annots$fullGenome, label = 'theRegion') |> 
   count(annotation_first, annotation_second) |> 
   as.data.frame() |> 
   dplyr::arrange(desc(n))

annotate(GM12878_HiCCUPS, annots$codingGenome, label = 'theRegion') |> 
   count(annotation_first, annotation_second) |> 
   as.data.frame() |> 
   dplyr::arrange(desc(n))
}
