% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginteractions-construct.R
\name{as_ginteractions}
\alias{as_ginteractions}
\title{Construct a GInteractions object from a tibble, DataFrame or data.frame}
\usage{
as_ginteractions(.data, ..., keep_mcols = TRUE)
}
\arguments{
\item{.data}{a \code{\link[=data.frame]{data.frame()}}, \code{\link[=DataFrame]{DataFrame()}} or \code{tibble()} to
construct a GInteractions object from.}

\item{...}{Optional named arguments specifying which the columns in .data
containin the core components a GInteractions object.}

\item{keep_mcols}{place the remaining columns into the metadata columns slot
(default=TRUE).}
}
\value{
a GInteractions object.
}
\description{
The \code{as_ginteractions} function looks for column names in
.data called seqnames\{1,2\}, start\{1,2\}, end\{1,2\}, and strand\{1,2\}
in order to construct a GInteractions object.
By default other columns in .data are placed into the mcols (metadata
columns) slot of the returned object.
}
\examples{
####################################################################
# 1. GInteractions from standard pairs files imported into a data.frame
####################################################################

pairs <- read.table(text = "
EAS139:136:FC706VJ:2:2104:23462:197393 chr1 10000 chr1 20000 + +
EAS139:136:FC706VJ:2:8762:23765:128766 chr1 50000 chr1 70000 + +
EAS139:136:FC706VJ:2:2342:15343:9863 chr1 60000 chr2 10000 + +
EAS139:136:FC706VJ:2:1286:25:275154 chr1 30000 chr3 40000 + -", 
col.names = c("readID", "chr1", "pos1", "chr2", "pos2", "strand1", "strand2"))
pairs |> 
  as_ginteractions(
    seqnames1 = chr1, start1 = pos1, end1 = pos1, 
    seqnames2 = chr2, start2 = pos2, end2 = pos2
  )

####################################################################
# 2. GInteractions from bedpe files imported into a data.frame
####################################################################

bedpe <- read.table(text = "
chr1 100 200 chr1 5000 5100 bedpe_example1 30 + -
chr1 1000 5000 chr1 3000 3800 bedpe_example2 100 + -",
col.names = c(
  "chrom1", "start1", "end1", 
  "chrom2", "start2", "end2", "name", "score", "strand1", "strand2"))
bedpe |> 
  as_ginteractions(seqnames1 = chrom1, seqnames2 = chrom2)

####################################################################
# 3. GInteractions from data.frame with extra fields
####################################################################

df <- read.table(text = "
chr1 100 200 chr1 5000 5100
chr1 1000 5000 chr1 3000 3800",
col.names = c("chr1", "start1", "end1", "chr2", "start2", "end2"))
df |> 
  as_ginteractions(seqnames1 = chr1, seqnames2 = chr2)

df <- read.table(text = "
chr1 100 200 chr1 5000 5100
chr1 1000 5000 chr1 3000 3800",
col.names = c("chr1", "start1", "end1", "chr2", "start2", "end2"))
df |> 
  as_ginteractions(seqnames1 = chr1, seqnames2 = chr2, strand1 = '+', strand2 = '-')

data.frame(type = "cis", count = 3) |> 
  as_ginteractions(
    seqnames1 = 'chr1', start1 = 1, end1 = 10,
    seqnames2 = 'chr1', start2 = 40, end2 = 50
  )
}
\seealso{
\code{\link[InteractionSet:GInteractions-class]{InteractionSet::GInteractions()}}
}
