% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginteractions-getters.R
\name{seqnames1}
\alias{seqnames1}
\alias{seqnames2}
\alias{start1}
\alias{start2}
\alias{end1}
\alias{end2}
\alias{width1}
\alias{width2}
\alias{strand1}
\alias{strand2}
\alias{seqnames1,GInteractions-method}
\alias{seqnames2,GInteractions-method}
\alias{start1,GInteractions-method}
\alias{start2,GInteractions-method}
\alias{end1,GInteractions-method}
\alias{end2,GInteractions-method}
\alias{width1,GInteractions-method}
\alias{width2,GInteractions-method}
\alias{strand1,GInteractions-method}
\alias{strand2,GInteractions-method}
\title{Internal GInteractions getters}
\usage{
seqnames1(x)

seqnames2(x)

start1(x)

start2(x)

end1(x)

end2(x)

width1(x)

width2(x)

strand1(x)

strand2(x)

\S4method{seqnames1}{GInteractions}(x)

\S4method{seqnames2}{GInteractions}(x)

\S4method{start1}{GInteractions}(x)

\S4method{start2}{GInteractions}(x)

\S4method{end1}{GInteractions}(x)

\S4method{end2}{GInteractions}(x)

\S4method{width1}{GInteractions}(x)

\S4method{width2}{GInteractions}(x)

\S4method{strand1}{GInteractions}(x)

\S4method{strand2}{GInteractions}(x)
}
\description{
Internal GInteractions getters
}
