% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyinteractions.R
\docType{package}
\name{plyinteractions-package}
\alias{plyinteractions}
\alias{plyinteractions-package}
\title{plyinteractions: a grammar of data manipulation for genomic interactions}
\description{
plyinteractions is a dplyr-like API to the
GInteractions infrastructure in Bioconductor.
}
\details{
plyinteractions provides a consistent interface for importing and
wrangling genomic interactions from a variety of sources. The package defines a
grammar of genomic interactions manipulation through a set of verbs. These verbs
can be used to construct human-readable analysis pipelines based on
\code{GInteractions}
\itemize{
\item Group genomic interactions with \code{group_by};
\item Summarize grouped genomic interactions with \code{summarize};
\item Tally/count grouped genomic interactions with \code{tally} and \code{count};
\item Modify genomic interactions with \code{mutate};
\item Subset genomic interactions with \code{filter} using
\href{https://rlang.r-lib.org/reference/args_data_masking.html}{\verb{<data-masking>}}
and logical expressions;
\item Pick out any columns from the associated metadata with \code{select}
using \href{https://dplyr.tidyverse.org/reference/dplyr_tidy_select.html}{\verb{<tidy-select>} arguments};
\item Subset using indices with \code{slice};
\item Order genomic interactions with \code{arrange} using categorical/numerical
variables.

For more details on the features of plyinteractions, read the vignette:
\code{browseVignettes(package = "plyinteractions")}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/js2264/plyinteractions}
  \item Report bugs at \url{https://github.com/js2264/plyinteractions/issues}
}

}
\author{
\strong{Maintainer}: Jacques Serizay \email{jacquesserizay@gmail.com}

}
