% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-dplyr.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{slice}
\alias{select}
\alias{filter}
\alias{arrange}
\title{Objects exported from other packages}
\value{
Depending on the re-exported function
}
\examples{
1 + 1
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}}
}}

